+++ app/src/main/res/layout/activity_main.xml
@@ -53,4 +53,13 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/tmp_user_type_button" />
 
+    <Button
+        android:id="@+id/tmp_login_button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        app:layout_constraintTop_toBottomOf="@+id/tmp_review"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        android:text="로그인 뷰로 임시 이동 버튼"/>
+
 </androidx.constraintlayout.widget.ConstraintLayout>
