+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,7 @@
 <resources>
     <string name="app_name">fooddeuk-android</string>
 
-    <!-- 밑줄 처리 -->
+    <!-- 밑줄 처리가 필요한 string -->
     <string name="terms_of_service_show_text"><u>보기</u></string>
+    <string name="go_sign_up"><u>회원가입 하기</u></string>
 </resources>
